feat: text capture record

+++ src/domain/essay/application/usecases/process-extract-text-from-image-usecase.ts
@@ -3,20 +3,26 @@ import { TextCaptureRecordRepository } from "../repositories/text-capture-record
 import { AiTemplateService } from "../../enterprise/services/ai-template-service";
 import { AiGateway } from "../gateways/ai-gateway";
 import { sleep } from "@/core/utils/sleep";
+import { Storage } from "../storage/storage";
+import { FilenameService } from "../../enterprise/services/filename-service";
 
 export class ProcessExtractTextFromImageUsecase {
   constructor(
     private textCaptureRecordRepository: TextCaptureRecordRepository,
-    private aiGateway: AiGateway
+    private aiGateway: AiGateway,
+    private storage: Storage
   ) {}
 
   async execute(props: Input): Promise<Output> {
     console.log(`Start extract text: `, props)
     await sleep(500)
     const textCaptureRecord = await this.textCaptureRecordRepository.getById(props.id)
-    if (!textCaptureRecord) throw new Error('TextCaptureRecord not found')
+    if (!textCaptureRecord || !textCaptureRecord.photoPath) throw new Error('TextCaptureRecord not found')
     const worker = await createWorker('eng');
-    const { data: { text: content }} = await worker.recognize(props.image_buffer);
+    const { data: { text: content }} = await worker.recognize(
+      FilenameService.toS3Url(textCaptureRecord.photoPath)
+    );
+    console.log(content)
     await worker.terminate();
     const question = AiTemplateService.getAdjustExtractedTextQuestion({ content });
     const answer = await this.aiGateway.adjustExtractedText(question);
@@ -28,8 +34,7 @@ export class ProcessExtractTextFromImageUsecase {
 }
 
 export type Input = {
-  id: string,
-  image_buffer: Buffer
+  id: string
 }
 
 export type Output = {}
